warpath/pyproject.toml
[project]
name = "warpath"
version = "0.1.0"
description = "Leveraged perpetual futures trading bot on Drift Protocol (Solana)"
requires-python = ">=3.11,<3.13"

dependencies = [
    "driftpy>=0.8.89",
    "solders>=0.21.0",
    "solana>=0.34.0",
    "anchorpy>=0.20.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "aiohttp>=3.9.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
backtest = [
    "matplotlib>=3.8.0",
]

[project.scripts]
warpath = "warpath.main:main"
